macos: remember what not to create in homedir

+++ macos-homedir.sh
@@ -8,6 +8,7 @@ if [ "$(uname)" != "Darwin" ]; then
 fi
 
 mkdir -p "$HOME/.shell-completion-local"
+mkdir -p "$HOME/.local/dotfiles"
 mkdir -p "$HOME/opt/bin"
 mkdir -p "$HOME/opt/lib"
 mkdir -p "$HOME/opt/sbin"
@@ -18,7 +19,7 @@ mkdir -p "$HOME/tmp"
 rm -rf "$HOME/.zoomus"
 touch "$HOME/.zoomus"
 
-if [ ! -d "$HOME/code" ] || [ ! -d "$HOME/env" ] || [ ! -d "$HOME/go" ]; then
+if { [ ! -d "$HOME/code" ] || [ ! -d "$HOME/env" ] || [ ! -d "$HOME/go" ] ;} && [ ! -e "$HOME/.local/dotfiles/no-home-dev-workflow-dirs" ]; then
   echo ""
   echo "Create dev workflow directories? (y/N)"
   read -r response
@@ -26,6 +27,8 @@ if [ ! -d "$HOME/code" ] || [ ! -d "$HOME/env" ] || [ ! -d "$HOME/go" ]; then
     mkdir -p "$HOME/code"
     mkdir -p "$HOME/go/bin"
     mkdir -p "$HOME/go/src"
+  else
+    touch "$HOME/.local/dotfiles/no-home-dev-workflow-dirs"
   fi
 fi
 
@@ -60,25 +63,29 @@ if [ ! -L "$HOME/Downloads/iCloud" ]; then
   ln -s "$HOME/Library/Mobile Documents/com~apple~CloudDocs/Downloads" "$HOME/Downloads/iCloud"
 fi
 
-if [ ! -L "$HOME/Books and Articles" ]; then
+if [ ! -L "$HOME/Books and Articles" ] && [ ! -e "$HOME/.local/dotfiles/no-home-booksandarticles-dir" ] ; then
   echo ""
   echo "Create link to Books & Articles in iCloud Drive? (y/N)"
   read -r response
   if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
     ln -s "$HOME/Library/Mobile Documents/com~apple~CloudDocs/Library" "$HOME/Books and Articles"
+  else
+    touch "$HOME/.local/dotfiles/no-home-booksandarticles-dir"
   fi
 fi
 
 if ! diff -r "$HOME/Library/Mobile Documents/com~apple~CloudDocs/Documents/" "$HOME/Documents" >/dev/null ; then
-  if [ ! -L "$HOME/Desktop/iCloud" ] || [ ! -L "$HOME/Documents/iCloud" ]; then
+  if { [ ! -L "$HOME/Desktop/iCloud" ] || [ ! -L "$HOME/Documents/iCloud" ] ;} && [ ! -e "$HOME/.local/dotfiles/no-home-icloud-links" ]; then
     echo ""
     echo "Desktop/Documents in iCloud appears to be disabled."
-    echo "Create links from Desktop/Documents to iCloud Drive (for WORK computer)? (y/N)"
-    echo "(eg. ~/Documents/Dropbox, etc.)"
+    echo "Create links from Desktop/Documents to iCloud Drive? (y/N)"
+    echo "(eg. ~/Documents/iCloud, etc. Mainly intended for work computers.)"
     read -r response
     if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
       ln -s "$HOME/Library/Mobile Documents/com~apple~CloudDocs/Desktop" "$HOME/Desktop/iCloud"
       ln -s "$HOME/Library/Mobile Documents/com~apple~CloudDocs/Documents" "$HOME/Documents/iCloud"
+    else
+      touch "$HOME/.local/dotfiles/no-home-icloud-links"
     fi
   fi
 fi
